Create Problem1.py
My initial conjecture for this problem was to take each number, subtract it from the k value and search for the difference inside of the list. However, this would result in a O(n2). I then thought more and decided to use a hash table which is a dictionary in python. I would attempt to look up the given number n in the dictionary, if it is not there then I will add a key to the dictionary of k - n and store the value as True. My initial thought was to simply return True if it found 2 numbers but I decided it may want to know what those numbers were, subsequently I decided to print the numbers out if the key was found in the dictionary. If there were not two numbers then the function will tell the user “No Matches found.” This solution runs in O(n), when we assume that the algorithm to look up keys is O(1).