Create MaxProduct.py -- Problem #69
This problem was from DailyCodingProblems, it successfully implements the task needed which was find the highest product of any 3 numbers. To solve this problem I decided to sort the list, and multiply the lowest 2 lowest negative numbers and the highest positive number; and compare it to multiplying the 3 highest numbers, and returning the greater of the two.